Refactor CompressionJobTable use <Panel/>
